+++ plugins/org.apache.cordova.file/www/FileSystem.js
@@ -32,7 +32,17 @@ var FileSystem = function(name, root) {
     this.name = name || null;
     if (root) {
         this.root = new DirectoryEntry(root.name, root.fullPath, this);
+    } else {
+        this.root = new DirectoryEntry(this.name, '/', this);
     }
+};
+
+FileSystem.prototype.__format__ = function(fullPath) {
+    return fullPath;
+};
+
+FileSystem.prototype.toJSON = function() {
+    return "<FileSystem: " + this.name + ">";
 };
 
 module.exports = FileSystem;