+++ plugins/org.apache.cordova.media-capture/www/MediaFile.js
@@ -32,7 +32,7 @@ var utils = require('cordova/utils'),
  * lastModifiedDate {Date} last modified date
  * size {Number} size of the file in bytes
  */
-var MediaFile = function(name, fullPath, type, lastModifiedDate, size){
+var MediaFile = function(name, localURL, type, lastModifiedDate, size){
     MediaFile.__super__.constructor.apply(this, arguments);
 };
 
@@ -48,7 +48,7 @@ MediaFile.prototype.getFormatData = function(successCallback, errorCallback) {
     if (typeof this.fullPath === "undefined" || this.fullPath === null) {
         errorCallback(new CaptureError(CaptureError.CAPTURE_INVALID_ARGUMENT));
     } else {
-        exec(successCallback, errorCallback, "Capture", "getFormatData", [this.fullPath, this.type]);
+        exec(successCallback, errorCallback, "Capture", "getFormatData", [this.localURL, this.type]);
     }
 };